requirements.txt
aiogram>=3.4.1,<4.0.0
SQLAlchemy>=2.0.29,<3.0.0
pydantic>=2.7.1,<3.0.0
python-dotenv>=1.0.1,<2.0.0
google-generativeai>=0.7.2,<1.0.0
Pillow>=10.3.0,<11.0.0